1.新增网址导入导出功能; 2.新增网址共享功能; 3.新增网址共享热度排序功能; 4.新增个人主页; 5.新增站内共享网址搜索;

+++ src/main/java/com/lyj/service/URLService.java
@@ -10,6 +10,7 @@ import com.lyj.model.Result;
 import com.lyj.model.URL;
 import com.lyj.model.User;
 import com.lyj.util.PublicVar;
+import com.lyj.util.RedisUtil;
 import com.lyj.util.ResultUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
@@ -19,7 +20,6 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -40,6 +40,9 @@ public class URLService {
 
     @Autowired
     HotUrlService hotUrlService;
+
+    @Autowired
+    User_HotUrlService user_hotUrlService;
 
     //在批量插入时使用（如果label相同有存在，则认为是重复的网址）
     public boolean isExistesUrl(URL url){
@@ -97,29 +100,81 @@ public class URLService {
         return i==1 ? true : false;
     }
 
-    public boolean deleteUrl(Integer id) {
+    public void deleteUrl(int id,boolean isShare,String userName) {
         int i = urlDao.deleteUrlByUrlId(id);
-        return i==1 ? true : false;
+        if(i==0){
+            throw new MessageException("删除失败!");
+        }
+
+        //如果这个要删除的网站是分享状态
+        if(isShare){
+            //先将数据库中的hotUrl表中的数据也删除
+            hotUrlService.deleteHotUrlByUrlId(id);
+            //再将user_hotUrl表中的关于该网址数据的点赞记录也删除
+            int deleteNumber = user_hotUrlService.deleteUser_HotUrlByUrlId(id);
+
+            //将共享数据保存到redis中，保证1点中的时候将redis中的数据同步到数据库中
+            redisTemplate.executePipelined(new RedisCallback<String>() {
+                @Override
+                public String doInRedis(RedisConnection connection) throws DataAccessException {
+                    connection.hDel(PublicVar.urlClickNumber.getBytes(), RedisUtil.toByte(id));//删除urlClickNumber中的url
+                    connection.hDel(PublicVar.urlGoodNumber.getBytes(), RedisUtil.toByte(id));//删除urlGoodNumber中的url
+
+                    connection.zRem(PublicVar.urlScore.getBytes(),RedisUtil.toByte(id));//删除urlScore排序集合中的数据
+                    connection.zIncrBy(PublicVar.userShareScore.getBytes(),-1.0,RedisUtil.toByte(userName));//减少userShareScore用户分享的个数
+                    connection.zIncrBy(PublicVar.userGoodScore.getBytes(),-deleteNumber*1.0,RedisUtil.toByte(userName));//减少userGoodScore用户点赞的个数
+                    return null;
+                }
+            });
+        }
     }
 
-    public boolean deleteUrlByPid(Integer id) {
-        int i = urlDao.deleteUrlByPid(id);
-        return i>=1 ? true : false;
+    public void deleteUrlByPid(Integer id,String userName) {
+        List<Integer> ids = urlDao.getShareUrlIdsByPid(id);//获取这个pid下的已经分享的urlId集合
+        int i = urlDao.deleteUrlByPid(id);//删除对应的url
+        deleteUrlDataBatchByIdsUtil(redisTemplate,ids,userName);//调用工具类批量清除数据和缓存
     }
 
-    public boolean addUrl(URL url) {
+    public void addUrl(URL url,User sessionUser) {
         int i = urlDao.addUrl(url);
-        return i==1 ? true : false;
+        if(i==0){
+            throw new MessageException("网址添加失败!");
+        }
+
+        if(url.getIsShare()){
+            //将共享数据保存到redis中，保证1点中的时候将redis中的数据同步到数据库中
+            redisTemplate.executePipelined(new RedisCallback<String>() {
+                @Override
+                public String doInRedis(RedisConnection connection) throws DataAccessException {
+                    connection.zIncrBy(PublicVar.urlScore.getBytes(),0.0,RedisUtil.toByte(url.getId()));//urlScore中添加记录
+                    connection.zIncrBy(PublicVar.userShareScore.getBytes(),1.0,sessionUser.getUserName().getBytes());//增加userShareScore用户分享的个数
+                    return null;
+                }
+            });
+        }
     }
 
-    public boolean deleteUrlsInBatchesByIds(String id) {
+    public void deleteUrlsInBatchesByIds(String id,String isShare,String userName) {
         String[] split = id.split(",");
+        String[] split_1 = isShare.split(",");
         List<Integer> ids=new ArrayList();
+        List<Integer> isShareIds=new ArrayList();
         for(int i=0;i<split.length;i++){
             ids.add(Integer.valueOf(split[i]));
+            if(split_1[i].equals("true")){
+                isShareIds.add(Integer.valueOf(split[i]));
+            }
         }
-        int i = urlDao.deleteUrlsByIds_Batch(ids);
-        return i>=1 ? true : false;
+        if(ids.size()>0){
+            int i = urlDao.deleteUrlsByIds_Batch(ids);
+            if(i==0){
+                throw new MessageException("批量删除失败!");
+            }
+            //只删除已经分享的数据缓存
+            if(isShareIds.size()>0){
+                deleteUrlDataBatchByIdsUtil(redisTemplate,isShareIds,userName);//调用工具类批量清除数据和缓存
+            }
+        }
     }
 
     public boolean updateUrlsInBatchesByIds(String id, int pid,String pidName) {
@@ -128,17 +183,32 @@ public class URLService {
         for(int i=0;i<split.length;i++){
             ids.add(Integer.valueOf(split[i]));
         }
-        int i = urlDao.updateUrlsByIds_Batch(ids,pid,pidName);
+
+        int i=0;
+        if(ids.size()>0){
+            i = urlDao.updateUrlsByIds_Batch(ids,pid,pidName);
+        }
         return i>=1 ? true : false;
 
     }
 
-    public int deleteUrlByUserId(Integer userId) {
-        return urlDao.deleteUrlByUserId(userId);
+    public void deleteUrlByUserId(Integer userId,String userName) {
+        List<Integer> ids = urlDao.getShareUrlIdsByUserId(userId);
+        int i = urlDao.deleteUrlByPid(userId);//删除对应的url
+        if(i==0) throw new MessageException("用户的链接删除失败");
+
+        deleteUrlDataBatchByIdsUtil(redisTemplate,ids,userName);//调用工具类批量清除数据和缓存
     }
 
-    public int addUrlBatch(List<URL> list) {
-        return urlDao.addUrlBatch(list);
+    public int addUrlBatch(List<URL> list,String userName) {
+        if(list.size()>0){
+            int i= urlDao.addUrlBatch(list);
+            if(i==0) throw new MessageException("批量添加失败,添加数为0");
+            addUrlDataInUrlScoreBatchByIdsUtil(redisTemplate,list,userName);//在redis中添加缓存
+            return i;
+        }else{
+            return 0;
+        }
     }
 
     public Result getUrlsByUserId(int userId) {
@@ -157,34 +227,28 @@ public class URLService {
             //先将数据库中的hotUrl表中的数据也删除
             hotUrlService.deleteHotUrlByUrlId(url.getId());
             //再将user_hotUrl表中的关于该网址数据的点赞记录也删除
-            hotUrlService.deleteUser_HotUrlByUrlId(url.getId());
+            int deleteNumber = user_hotUrlService.deleteUser_HotUrlByUrlId(url.getId());
 
             //将共享数据保存到redis中，保证1点中的时候将redis中的数据同步到数据库中
             redisTemplate.executePipelined(new RedisCallback<String>() {
                 @Override
                 public String doInRedis(RedisConnection connection) throws DataAccessException {
-                    connection.hDel(String.valueOf(url.getId()).getBytes());//删除缓存的共享的url数据
-                    connection.zRem(PublicVar.hotUrlScore.getBytes(),String.valueOf(url.getId()).getBytes());//删除掉redis中进行排序的数据
-                    try {
-                        //将中文以utf-8的格式进行编码
-                        connection.zIncrBy(PublicVar.userShareScore.getBytes(),-1L,sessionUser.getUserName().getBytes("utf-8"));//减少redis中记录用户分享的个数
-                    } catch (UnsupportedEncodingException e) {
-                        e.printStackTrace();
-                    }
+                    connection.hDel(PublicVar.urlClickNumber.getBytes(), RedisUtil.toByte(url.getId()));//删除urlClickNumber中的url
+                    connection.hDel(PublicVar.urlGoodNumber.getBytes(), RedisUtil.toByte(url.getId()));//删除urlGoodNumber中的url
+
+                    connection.zRem(PublicVar.urlScore.getBytes(),RedisUtil.toByte(url.getId()));//删除urlScore排序集合中的数据
+                    connection.zIncrBy(PublicVar.userShareScore.getBytes(),-1.0,sessionUser.getUserName().getBytes());//减少userShareScore用户分享的个数
+                    connection.zIncrBy(PublicVar.userGoodScore.getBytes(),-deleteNumber*1.0,sessionUser.getUserName().getBytes());//减少v用户点赞的个数
                     return null;
                 }
             });
-//            Boolean delete = redisTemplate.delete(String.valueOf(url.getId()));//删除缓存的共享的url数据
-//            Long num=redisTemplate.opsForZSet().remove(PublicVar.hotUrlScore,url.getId());//删除掉redis中进行排序的数据
-//            redisTemplate.opsForZSet().incrementScore(PublicVar.userShareScore,sessionUser.getUserName(),-1L);//减少redis中记录用户分享的个数
         }else{//如果选择共享
             //将共享数据保存到redis中，保证1点中的时候将redis中的数据同步到数据库中
             redisTemplate.executePipelined(new RedisCallback<String>() {
                 @Override
                 public String doInRedis(RedisConnection connection) throws DataAccessException {
-                    connection.hSetNX(String.valueOf(url.getId()).getBytes(), "url".getBytes(), JSON.toJSONString(url).getBytes());//保存url信息
-                    connection.zIncrBy(PublicVar.hotUrlScore.getBytes(),0,String.valueOf(url.getId()).getBytes());//在分数表中创建记录
-                    connection.zIncrBy(PublicVar.userShareScore.getBytes(),1,sessionUser.getUserName().getBytes());//增加redis中记录用户分享的个数
+                    connection.zIncrBy(PublicVar.urlScore.getBytes(),0.0,RedisUtil.toByte(url.getId()));//urlScore添加记录
+                    connection.zIncrBy(PublicVar.userShareScore.getBytes(),1.0,sessionUser.getUserName().getBytes());//增加userShareScore用户分享的个数
                     return null;
                 }
             });
@@ -193,7 +257,45 @@ public class URLService {
     }
 
     public List<URL> getUrlsByIdBatch(List list) {
-        return urlDao.getUrlsByIdBatch(list);
+        if(list.size()>0){
+            return urlDao.getUrlsByIdBatch(list);
+        }else{
+            return new ArrayList<>();
+        }
+    }
+
+
+    //批量删除网址后清除数据库和缓存的工具
+    public void deleteUrlDataBatchByIdsUtil(RedisTemplate redisTemplate,List<Integer> ids,String userName){
+        int deleteNumber = hotUrlService.deleteHotUrlByUrlIdBatch(ids);
+        int goodNumber = user_hotUrlService.deleteUser_HotUrlByUrlIdBatch(ids);//删除并返回这个网址所点赞的个数
+
+        redisTemplate.executePipelined(new RedisCallback<String>() {
+            @Override
+            public String doInRedis(RedisConnection connection) throws DataAccessException {
+                connection.hDel(PublicVar.urlClickNumber.getBytes(), RedisUtil.toByteArray(ids));//删除urlClickNumber中的url
+                connection.hDel(PublicVar.urlGoodNumber.getBytes(), RedisUtil.toByteArray(ids));//删除urlGoodNumber中的url
+
+                connection.zRem(PublicVar.urlScore.getBytes(),RedisUtil.toByteArray(ids));//删除urlScore排序集合中的数据
+                connection.zIncrBy(PublicVar.userShareScore.getBytes(),-ids.size()*1.0,RedisUtil.toByte(userName));//减少userShareScore用户分享的个数
+                connection.zIncrBy(PublicVar.userGoodScore.getBytes(), -goodNumber*1.0,RedisUtil.toByte(userName));//减少userGoodScore用户点赞的个数
+                return null;
+            }
+        });
+    }
+
+    //批量添加网址分享记录在urlScore中
+    public void addUrlDataInUrlScoreBatchByIdsUtil(RedisTemplate redisTemplate,List<URL> ids,String userName){
+        redisTemplate.executePipelined(new RedisCallback<String>() {
+            @Override
+            public String doInRedis(RedisConnection connection) throws DataAccessException {
+                for(int i=0;i<ids.size();i++){
+                    connection.zIncrBy(PublicVar.urlScore.getBytes(),0.0, RedisUtil.toByte(ids.get(i).getId()));//在urlScore中添加记录
+                }
+                connection.zIncrBy(PublicVar.userShareScore.getBytes(),ids.size()*1.0,RedisUtil.toByte(userName));//在userShareScore中更新用户分享的个数
+                return null;
+            }
+        });
     }
 
 }
